Refatoracao EventBus
