+++ app/src/main/java/com/fulltime/foodex/helper/eventbus/ChangeEmpresa.java
@@ -1,0 +1,34 @@
+package com.fulltime.foodex.helper.eventbus;
+
+/*
+  FoodEx is a sales management application.
+  Copyright (C) 2019 Josue Lopes.
+  <p>
+  FoodEx is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+  <p>
+  FoodEx is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+  <p>
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import com.fulltime.foodex.model.Empresa;
+
+public class ChangeEmpresa {
+
+    private final Empresa empresa;
+
+    public ChangeEmpresa(Empresa empresa) {
+        this.empresa = empresa;
+    }
+
+    public Empresa getEmpresa() {
+        return empresa;
+    }
+}
