+++ app/src/main/java/com/fulltime/foodex/helper/eventbus/FirstCorporation.java
@@ -1,15 +1,15 @@
-package com.fulltime.foodex.helper.update;
+package com.fulltime.foodex.helper.eventbus;
 
 /*
   FoodEx is a sales management application.
   Copyright (C) 2019 Josue Lopes.
   <p>
-  FoodEx is free software: you can redistribute it and/or modify
+  This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
   (at your option) any later version.
   <p>
-  FoodEx is distributed in the hope that it will be useful,
+  This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.
@@ -18,19 +18,16 @@ package com.fulltime.foodex.helper.update;
   along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import com.fulltime.foodex.model.Cliente;
+import com.fulltime.foodex.model.Empresa;
 
-import java.util.List;
+public class FirstCorporation {
+    private final Empresa empresa;
 
-public class ListaCliente {
-
-    private final List<Cliente> clientes;
-
-    public ListaCliente(List<Cliente> clientes) {
-        this.clientes = clientes;
+    public FirstCorporation(Empresa empresa) {
+        this.empresa = empresa;
     }
 
-    public List<Cliente> getClientes() {
-        return clientes;
+    public Empresa getEmpresa() {
+        return this.empresa;
     }
 }
