+++ app/src/main/java/com/fulltime/foodex/helper/eventbus/ListaClienteVazia.java
@@ -1,4 +1,4 @@
-package com.fulltime.foodex.helper.update;
+package com.fulltime.foodex.helper.eventbus;
 
 /**
  * FoodEx is a sales management application.
